+++ SpeechResponder/CustomFunctions/BodyDetails.cs
@@ -4,6 +4,7 @@ using EddiDataDefinitions;
 using EddiSpeechResponder.ScriptResolverService;
 using JetBrains.Annotations;
 using System;
+using System.Linq;
 using System.Reflection;
 
 namespace EddiSpeechResponder.CustomFunctions
@@ -17,22 +18,36 @@ namespace EddiSpeechResponder.CustomFunctions
         public Type ReturnType => typeof( Body );
         public IFunction function => Function.CreateNativeMinMax( ( runtime, values, writer ) =>
         {
-            StarSystem system;
-            if (values.Count == 0)
+            Body body;
+            if ( values.Any() )
             {
-                system = EDDI.Instance.CurrentStarSystem;
-            }
-            else if (values.Count == 1 || string.IsNullOrEmpty(values[1].AsString) || (values.Count > 1 && values[1].AsString?.ToLowerInvariant() == EDDI.Instance.CurrentStarSystem?.systemname?.ToLowerInvariant()))
-            {
-                system = EDDI.Instance.CurrentStarSystem;
+                StarSystem system;
+                if ( values.Count < 2 || string.IsNullOrEmpty( values[ 1 ].AsString ) )
+                {
+                    // Current system
+                    system = EDDI.Instance.CurrentStarSystem;
+                }
+                else
+                {
+                    // Named system
+                    system = ulong.TryParse( values[ 1 ].AsString, out var systemAddress )
+                        ? EDDI.Instance.DataProvider.GetOrFetchStarSystem( systemAddress, true, false )
+                        : EDDI.Instance.DataProvider.GetOrFetchStarSystem( values[ 1 ].AsString, true, false );
+                }
+
+                body = long.TryParse( values[ 0 ].AsString, out var bodyID )
+                    ? system?.bodies?.Find( v => v.bodyId == bodyID )
+                    : system?.bodies?.Find( v =>
+                        v.bodyname?.ToLowerInvariant() == values[ 0 ].AsString?.ToLowerInvariant() );
             }
             else
             {
-                // Named system
-                system = EDDI.Instance.DataProvider.GetOrFetchStarSystem( values[1].AsString, true, false );
+                body = EDDI.Instance.CurrentStellarBody;
             }
-            var result = system?.bodies?.Find(v => v.bodyname?.ToLowerInvariant() == values[0].AsString?.ToLowerInvariant());
-            return result is null ? Value.EmptyMap : Value.FromReflection( result, BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic );
+
+            return body is null
+                ? Value.EmptyMap
+                : Value.FromReflection( body, BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic );
         }, 1, 2);
     }
 }